labels: simplified labelling of PRs.
Previously we adding labels, we actually edited issues with an array of all the labels
the issue should have. That meant we had to include both the pre existing labels, as well
as the labels we resolved by the filepaths changed.

Now with the upgrade of the github npm dependency, we got a new API for *adding new labels*
to issues. That means we don't have to fetch pre existing labels and so forth,
simplifying the process a whole lot.