labels: integration test for version label
This adds an integration test which verifies that the bot
tries to add version label (e.g. v6.x) when the github.com
webhook payload says the PR opened is targeted against the
v6.x-staging branch.

Detailed tests and variations are still done in unit tests,
this is just another level of verification which tests *a lot*
of the code involved in labelling, and especially the webhook
payload content.