jenkins: allow citgm builds to push Jenkins build updates
This changes allows citgm builds to push updates like we already do
for nodejs/node. That is done by making the Jenkins status update
endpoints more generic, resolving the repository name from an URL
parameter.

In practise this means the endpoints stay the same, but configurable:
- `/node/jenkins/start`
- `/citgm/jenkins/start`

Repository names are validated against a whitelist.