jenkins: POC for kicking off builds based on PR comments
This is the initial shot at triggering a Jenkins build when a repository
collaborator mentions the bot in a comment with the following content:

`@nodejs-github-bot run CI`

In addition the bot has to be explicitly configured with $ENV variables
per repo and related Jenkins build for this to be enabled.

At first we'll start with https://github.com/nodejs/citgm which has been
the biggest driver for getting this implemented. If that test run succeeds
we can enable it on other repos as we see fit in collaboration with their
corresponding working group.

Refs https://github.com/nodejs/github-bot/issues/82