fix(jenkins): ignore builds not related to pull requests
We have for a long time seen errors logged when Jenkins tells the bot
to report inline pull requests statuses, when the build in question is
not related to a pull request, but an ordinary branch like the "v8.x-staging"
branch.

In short the error raised by github.com said we couldn't fetch commits
related to pull request "v8.x-staging". It expects a pull request number
to be provided when fetching those commits.

These changes checks if the git reference provided by Jenkins does in
fact point to a pull request, if not the incoming request from Jenkins
gets a 400 Bad Request as response.

```
13:01:11.244 ERROR bot: Got error when retrieving GitHub commits for PR (req_id=78e986d9-6c52-485c-bac2-64c2797a7a40, pr=v8.x-staging, job=node-test-commit-aix, status=failure)
    err: {
      "code": "400",
      "status": "Bad Request",
      "message": "Invalid value for parameter 'number': v8.x-staging"
    }
```