fix(jenkins): ignore ended builds not related to pull requests
This is an improvement of a recent attemt to ignore any Jenkins build
status updates that was *not* related to pull requests.

While looking at the logs after the last fix was merged, it surprised
me that we still got those kinds of errors popping up:

```
22:24:02.994 ERROR bot: Got error when retrieving GitHub commits for PR (req_id=3741a210-6cec-11e8-a3a3-e194f9df710c, pr=master, job=node-test-commit-plinux, status=success)
    err: {
      "code": "400",
      "status": "Bad Request",
      "message": "Invalid value for parameter 'number': master"
    }
```

Looking closely on the info included in that error log, made it clear it
could be an issue on builds that had *ended*. Digging into the fix
that was recently pushed, I discovered that fix only applied to for statuses
Jenkins pushes to the bot when it *starts* builds.

Applying the previous fix to the bot endpoint that handles *ended* Jenkins
builds as well, should make these error logs disappear.

Refs https://github.com/nodejs/github-bot/pull/177