jenkins: wait until status has been pushed to GitHub before responding
These changes are primarily done so we can get more reliable integration
tests. Previously the incoming requests was responded to, before we
knew whether or not the Jenkins status has been successfully pushed to
GitHub.

That makes it challenging to know when we can assert what we want to
verify in integration tests, as we don't know when the entire operation
has finished. Although the previous approach did what it should in
practise when run in production, it should be just as important to have
reliable tests that we can rely on when doing changes going forward.