jenkins: update existing comment when CI comment has already been posted
The motivation behind these changes are to prevent flooding PRs with CI
comments from the bot whenever new node-test-pull-request jobs are
started. Instead if the bot has already posted a CI comment earlier,
it will update the last CI comment and change the URL in that comment
to the new job that got started.