jenkins: destructure build result after knowing it's not an error
We've seen some error logs in production suggesting that we get errors
back from Jenkins when triggering builds, as the result is
`null | undefined`. Destructuring fields out of that result causes
unchaught exceptions, meaning the bot process dies.

Therefore postponing the destructuring until we're confident the build
didn't result in an error.