chore: update github v12 > v13 due to vulnerability in https-proxy-agent
Since this is a major update of *the* most important dependency in this
project, namely the one we use to communicate to github with, it
deserves its own bump commit, making it easier to revert in case any
surprises are found.

Ideally we would jump to the latest and greatest version of this
dependency, which has even been renamed to `@octokit/rest`, but doing
so sadly breakes a lot of tests ATM. Will have to look into that later.

Refs https://www.npmjs.com/advisories/1184