feat: update bot comment when Jenkins finishes
Update the bot comment on a PR when the corresponding Jenkins build
finishes. The updated comment will have the build status, and if the
build failed, it will have a summary of failures.

To identify a comment corresponding to a PR, we leave placeholders on
the comment as HTML commnets, so those placeholders can be replaced
later.

Signed-off-by: Matheus Marchini <mmarchini@netflix.com>