chore: remove unused attempt backport script
This was a very neat part of the bot at one point, but ended up creating
some noise and got disabled.

It used to try backport of patches from created & updated PRs against
the `${version}.x-staging`s and as a result, added
`lts-watch-v${version}.x` or `dont-land-on-v${version}.x` when
applicable.

When it was operationale, it tried to apply patches against Node.js
versions: 4, 6 and 7. Those version are all end-of-life now.

For the bot to try backport of patches for newer versions of Node.js,
those versions would have to be added explicitly in the bot' source code.
That has obviously not happened, assuming the reason being it wasn't
as helpful afterall.