labels: disable labelling of PRs as it's replaced by GitHub Action
First step of removing labelling of PRs as they are opened. Only the
part actively listening for newly opened PRs gets removed now, with the
main goal of *stopping* @nodejs-github-bot from applying labels and
thereby leaving room for nodejs/node-pr-labeler GitHub Action to apply
those labels intead.

Still a lot of code that can be removed, like the actually logic of
resolving labels which has been extracted into the said GitHub Action,
but postponing that to a followup PR, when seeing the Action behaving
as expected -- if not, these changes might be reverted as a fallback.

Refs https://github.com/nodejs/node-pr-labeler